Fixes for missed data.

+++ models/src/main/scala/uk/gov/ons/bi/ingest/parsers/ImplicitHelpers.scala
@@ -22,9 +22,10 @@ object ImplicitHelpers {
   }
 
   implicit class TypedString(s: String) {
+
     def asDateTime = parseDate(s)
 
-    def asIntOpt = strOption(s).map(_.toInt)
+    def asIntOpt = strOption(s).map(_.toDouble.toInt)
 
     def asDateTimeOpt = {
       // FIXME: strOption(s).map(x => parseDate(x))
